Fixed a bug where item NetID's were sending as a Byte instead of Single, Todo TILEPLACEMENT

+++ Terraria_Server/NetMessage_Packets.cs
@@ -97,7 +97,7 @@ namespace Terraria_Server
 			
 			Byte (prefix);
 			
-			Byte (item.NetID);
+			Short (item.NetID);
 			
 			End ();
 		}
